% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequent_listened_artists.R
\name{frequent_listened_artists}
\alias{frequent_listened_artists}
\title{Generate a Frequent Listened Artists' Name plot based on Spotify data}
\usage{
frequent_listened_artists(
  category = "saved",
  data = spotifywRapped::saved_tracks,
  vibe = "neon",
  name = "untitled",
  saveto = getwd()
)
}
\arguments{
\item{category}{Character string indicating the category for the data. Must
be one of "long" or "saved". Default is "saved."}

\item{data}{Character string indicating the type of Spotify data to use.
Can be "long" for long-term top tracks or "saved" for saved tracks.}

\item{vibe}{Character string indicating the vibe of the word cloud.
Can be one of "bright", "neutral", "neon", or "soft".}

\item{name}{Character string indicating the name of the output file.}

\item{saveto}{Character string indicating the directory to save the output
file.}
}
\value{
A character string indicating the path to the saved word cloud image.
}
\description{
Generate a Frequent Listened Artists' Name plot based on Spotify data
}
