% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_artist_frequency.R
\name{my_artist_frequency}
\alias{my_artist_frequency}
\title{Artist Frequency Bar Graph}
\usage{
my_artist_frequency(
  dataset = data.frame(),
  category = "saved",
  vibe = "neon",
  name = "artist_frequency",
  saveto = getwd()
)
}
\arguments{
\item{dataset}{A data frame containing the user's saved or top tracks.
Defaults to package data.}

\item{category}{A character string specifying whether the tracks come from
the saved or top tracks.}

\item{vibe}{A character string specifying the vibe of the graphic. Must be
one of "bright", "dark", "neon", or "soft".}

\item{name}{A character string specifying the name of the graphic.}

\item{saveto}{A character string specifying the directory where the graphic
should be saved. Defaults to the current working directory.}
}
\value{
The file path of the .png file.
}
\description{
Create a bar graph demonstrating the top 10 artists, and how many songs of
each artist.
}
