% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_top_five.R
\name{my_top_five}
\alias{my_top_five}
\title{Visualize your top five artists or tracks}
\usage{
my_top_five(
  time,
  vibe,
  category = "artists",
  name = "my_top_five",
  saveto = getwd(),
  dataset = data.frame()
)
}
\arguments{
\item{time}{A character string specifying the time frame for the data. Must
be one of "long", "medium", or "short".}

\item{vibe}{A character string specifying the vibe of the graphic. Must be
one of "bright", "neutral", "neon", or "soft".}

\item{category}{A character string specifying whether the data is for artists
or
tracks. Must be one of "artists" or "songs".}

\item{name}{A character string specifying the name of the graphic.}

\item{saveto}{A character string specifying the directory where the graphic
should be saved. Defaults to the current working directory.}

\item{dataset}{A data frame containing the user's top artists or tracks.
Defaults to package data.}
}
\value{
The file path of the .png file.
}
\description{
This function creates a visualization of a user's top five artists or tracks
based on the vibe specified and whether the data is long-term, medium-term,
or short-term (all time, last 6 months, last 4 weeks, respectively).
The user specifies both the name and storage location of the graphic, which
is saved as a .png file.
}
