% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_wordcloud.R
\name{my_wordcloud}
\alias{my_wordcloud}
\title{Generate a Frequent Listened Artists' Name plot based on Spotify data}
\usage{
my_wordcloud(
  data = data.frame(),
  category = "long",
  vibe = "neon",
  name = "wordcloud",
  saveto = getwd()
)
}
\arguments{
\item{data}{Dataframe indicating the Spotify data to use. \
Defaults to blank data frame}

\item{category}{Character string indicating the category for the data. Must
be one of "long" or "saved". Default is "long."}

\item{vibe}{Character string indicating the vibe of the word cloud.
Can be one of "bright", "neutral", "neon", or "soft".}

\item{name}{Character string indicating the name of the output file.}

\item{saveto}{Character string indicating the directory to save the output
file.}
}
\value{
A character string indicating the path to the saved word cloud image.
}
\description{
Generate a Frequent Listened Artists' Name plot based on Spotify data
}
