% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saved_tracks}
\alias{saved_tracks}
\title{Saved Tracks - All time}
\format{
\subsection{\code{saved_tracks}}{

A dataframe with 765 rows and 33 columns:

\describe{
\item{track.artists}{List of dataframes containing artist(s) name, ID, and
photo URLs}
\item{track.id}{Track's unique Spotify ID.
Can be used to access track audio. Character.}
\item{track.name}{Track title. Character.}
\item{artist}{Principal artist on the track. Not natively pulled from
API--reference \code{DATASET.R} to view code to add this column. Character.}
\item{energy}{Audio feature called 'energy' of the track. Not natively
pulled from API--reference \code{DATASET.R} to view code to add this column.
Numeric.}
\item{valence}{Audio feature called 'acousticness' of the track. Not
natively pulled from API--reference \code{DATASET.R} to view code to add this
column. Numeric.}
...
}
@source \url{https://developer.spotify.com/documentation/web-api/reference/get-users-saved-tracks}
}
}
\usage{
saved_tracks
}
\description{
All of Jared Hileman's saved Spotify tracks.
}
\keyword{datasets}
