% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifywRapped-package.R
\docType{package}
\name{spotifywRapped-package}
\alias{spotifywRapped}
\alias{spotifywRapped-package}
\title{spotifywRapped: Creates Beautiful Visualizations of Spotify Premium User Listening Data}
\description{
The spotifywRapped package gives Spotify premium users control over how and when they get to view their listening history. This package depends on users having a Spotify premium account and data pulled from the Spotify API (perhaps using the spotifyr package). The spotifywRapped package lets the user specify a particular theme, or 'vibe', for their graphic of choice. A user has the option to choose between several types of graphics, like wordclouds visualizing their favorite artists, or an energy/ valence plot describing their listening habits. The user can then save these graphics wherever they choose!
}
\author{
\strong{Maintainer}: Alyssa Brock \email{brockflocka@gmail.com}

Authors:
\itemize{
  \item Jared Hileman \email{jared.hileman77@gmail.com}
  \item Lauren Rose \email{a02357321@usu.edu}
  \item Saichand Thota \email{s.thota@usu.edu}
}

}
\keyword{internal}
