% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{top_artists_shortterm}
\alias{top_artists_shortterm}
\title{Top Artists - Shortterm}
\format{
\subsection{\code{top_artists_shortterm}}{

A data frame with 4 rows and 11 columns:
\describe{
\item{name}{Artist's name, character}
\item{id}{Artist's unique Spotify ID, character}
\item{images}{Artist's photo URLs as a dataframe. Different size options
available}
\item{genres}{Artist's tracks' genres, stored as list}
...
}
}
}
\source{
\url{https://developer.spotify.com/documentation/web-api/reference/get-users-top-artists-and-tracks}
}
\usage{
top_artists_shortterm
}
\description{
Top / most-played of Jared Hileman's Spotify artists pulled from
approximately March 26th, 2024 - April 28th, 2024 (around 4 weeks).
}
\keyword{datasets}
