% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{top_tracks_longterm}
\alias{top_tracks_longterm}
\title{Top Tracks - Longterm}
\format{
\subsection{\code{top_tracks_longterm}}{

A data frame with 1000 rows and 29 columns:
\describe{
\item{artists}{Dataframe containing artist(s) name, ID, and photo URLs}
\item{available_markets}{Countries in which song is availabe}
\item{id}{Track's unique Spotify ID. Can be used to access track audio
features. Character}
\item{name}{Track title}
\item{album.images}{Album cover image URL(s) to which the track belongs}
...
}
}
}
\source{
\url{https://developer.spotify.com/documentation/web-api/reference/get-users-top-artists-and-tracks}
}
\usage{
top_tracks_longterm
}
\description{
Top / most-played of Jared Hileman's Spotify tracks pulled from approximately
April 23rd, 2023 - April 23th, 2024 (around 1 year -- limit: 1000
observations).
}
\keyword{datasets}
