% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_energy_valence.R, R/plot_energyvalence.R
\name{my_energy_valence}
\alias{my_energy_valence}
\title{Create an energy valence plot}
\usage{
my_energy_valence(
  data = spotifywRapped::saved_tracks,
  vibe = "neon",
  name = "untitled",
  saveto = getwd()
)

my_energy_valence(
  data = spotifywRapped::saved_tracks,
  vibe = "neon",
  name = "untitled",
  saveto = getwd()
)
}
\arguments{
\item{data}{a data frame containing songs and values between 0 and 1 for
energy and valence.  Defaults to relevant data from data raw folder.}

\item{vibe}{string that specifies the background colors and graphics.
Either "soft", "neon", "neutral", or "bright".}

\item{name}{string to be used as the name for the output file.
Defaults to "untitled".}

\item{saveto}{string to be used as the directory for the output file.
Defaults to working directory.}
}
\value{
the file path to the generated .png file

the file path to the generated .png file
}
\description{
\code{my_energy_valence} generates a PNG file that contains background colors and
graphics as specified by \code{vibe}.  It also displays a scatter plot of energy
vs. valence using the values in \code{data} and calculates which quadrant of the
graph contains the most points.

\code{my_energy_valence} generates a PNG file that contains background colors and
graphics as specified by \code{vibe}.  It also displays a scatter plot of energy
vs. valence using the values in \code{data} and calculates which quadrant of the
graph contains the most points.
}
